Move classes to dedicated files, remove unused ones
