+++ constants.h
@@ -1,0 +1,14 @@
+#ifndef _CONSTANTS_H_
+#define _CONSTANTS_H_
+
+#define PI 3.14159265358979
+#define ME 9.10938188e-31    //kg
+#define EL 1.60217646e-19    //coulombs
+#define HPLANCK 6.626068e-34 //m2 kg / s
+#define HBAR 1.05457148e-34  //m2 kg / s
+#define KBOL 8.6173324e-5    //eV K-1
+#define THREADS 56
+
+//1 electron volt = 1.60217646e-19 joule
+
+#endif
